insight_face/requirements.txt
insightface
opencv-python
matplotlib
requests
onnxruntime-gpu
numpy==1.20
python-multipart
fastapi
uvicorn
pillow
pydantic